pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drf-dynamic-fields"
version = "0.0.1"
authors = [
  { name="Erik Polzin", email="eriktpol@gmail.com" },
]
description = "Dynamic Fields for Django Rest Framework"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    'Django>=5.0.0',
    'djangorestframework>=3.0.0',
]

[project.optional-dependencies]
build = ["build", "twine"]
dev = ["pip-tools", "bumpver", "pytest", "pytest-django"]

[project.urls]
Homepage = "https://github.com/ErikPolzin/drf-dynamic-fields"
Issues = "https://github.com/ErikPolzin/drf-dynamic-fields/issues"

[tool.bumpver]
current_version = "2024.1001-alpha"
version_pattern = "YYYY.BUILD[-TAG]"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
pre_commit_hook = ""
post_commit_hook = ""
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
]
"README.md" = [
    "{version}",
    "{pep440_version}",
]

